+++ tests/Gate.IO.Api.Tests/GlobalUsings.cs
@@ -1,5 +1,6 @@
 global using Gate.IO.Api;
 global using Gate.IO.Api.Alpha;
+global using Gate.IO.Api.Wallet;
 global using Newtonsoft.Json;
 global using Newtonsoft.Json.Linq;
 global using Xunit;